Fix constants, move helper to prevent circular dep
